pyproject.toml
[tool.poetry]
name = "technova-attrition-analysis"
version = "0.1.0"
description = "Projet 4 - Analyse de l'attrition des employés chez TechNova Partners : exploration des données RH, modélisation prédictive et identification des facteurs de départ."
authors = ["barbara di battista"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

# === Outils généraux ===
requests = ">=2.32.5,<3.0.0"
python-dotenv = ">=1.1.1,<2.0.0"
tqdm = ">=4.67.1,<5.0.0"

# === Data manipulation ===
pandas = ">=2.3.3,<3.0.0"
numpy = ">=2.3.3,<3.0.0"

# === Visualisation ===
matplotlib = ">=3.10.6,<4.0.0"
seaborn = "^0.13.2"
pillow = ">=11.3.0,<12.0.0"

# === Jupyter / Notebooks ===
jupyter = "^1.1.1"
ipykernel = ">=6.30.1,<7.0.0"

# === Machine Learning - Base ===
scikit-learn = ">=1.7.2,<2.0.0"

# === Machine Learning - Modèles avancés ===
xgboost = "^3.1.2"
catboost = "^1.2.8"

# === Gestion du déséquilibre des classes ===
imbalanced-learn = "^0.14.0"

# === Interprétabilité des modèles ===
shap = "^0.50.0"
sqlalchemy = "^2.0.46"
psycopg2-binary = "^2.9.11"
pydantic = "^2.12.5"
fastapi = "^0.133.1"
uvicorn = {extras = ["standard"], version = "^0.41.0"}
joblib = "^1.5.3"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.22"
bcrypt = "4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"